progress is now in double

+++ qgrep/src/output.hpp
@@ -11,7 +11,7 @@ public:
 	virtual void print(const char* message, ...) = 0;
 	virtual void error(const char* message, ...) = 0;
 
-	virtual void progress(int percentage) {}
+	virtual void progress(double percentage) {}
 	virtual bool isStopped() { return false; }
 
 	virtual bool isTTY() { return false; }
